splited Home page on components
