+++ react/src/components/MyCitiesList/MyCitiesList.css
@@ -1,0 +1,14 @@
+.city-name {
+    font-weight: bold;
+    font-size: 1.2rem;
+}
+
+.city-card-mini {
+    border: 1px solid transparent;
+    border-radius: 0.25rem;
+}
+
+.city-card-mini:hover {
+    border: 1px solid lightskyblue;
+    cursor: pointer;
+}
