--- react/src/styles/Home.css
@@ -1,40 +1,0 @@
-.add-city-wrapper {
-    padding: 0 !important;
-}
-
-.city-name {
-    font-weight: bold;
-    font-size: 1.2rem;
-}
-
-.city-card-mini {
-    border: 1px solid transparent;
-    border-radius: 0.25rem;
-}
-
-.city-card-mini:hover {
-    border: 1px solid lightskyblue;
-    cursor: pointer;
-}
-
-.autocomplete-wrapper {
-    position: absolute;
-    background: #fff;
-    z-index: 1000;
-    width: 560px;
-    border: 1px solid lightgray;
-    border-radius: 5px;
-    margin-top: -0.75rem;
-    line-height: 1.25;
-}
-
-.autocomplete-item {
-    position: relative;
-    height: 40px;
-    cursor: pointer;
-    padding: 12px 12px 12px 10px;
-}
-
-.autocomplete-item:hover {
-    background-color: lightsteelblue;
-}
